Move InternalFetchAwareBuildController to org.gradle.tooling.internal.protocol package
